+++ backend/api/src/routes/modules/assignments/submissions/post.rs
@@ -117,8 +117,8 @@ fn within_late_window(
 }
 
 /// Returns (adjusted_earned, capped_to_opt)
-fn cap_late_earned(earned: i64, total: i64, max_percent: u32) -> (i64, Option<i64>) {
-    let cap = ((max_percent as i128) * (total as i128) / 100) as i64;
+fn cap_late_earned(earned: f32, total: f32, max_percent: f32) -> (f32, Option<f32>) {
+    let cap = max_percent * total / 100.0;
     if earned > cap {
         (cap, Some(cap))
     } else {
@@ -128,8 +128,8 @@ fn cap_late_earned(earned: i64, total: i64, max_percent: u32) -> (i64, Option<i6
 
 #[derive(Serialize)]
 struct WsMark {
-    earned: i64,
-    total: i64,
+    earned: f32,
+    total: f32,
 }
 
 async fn emit_submission_status(
@@ -268,7 +268,7 @@ pub async fn check_disallowed_code(
                 assignment_id,
                 user_id,
                 attempt,
-                0,
+                0.0,
                 allocator.total_value,
                 is_practice,
                 file_name,
@@ -311,7 +311,7 @@ pub async fn check_disallowed_code(
                 created_at: now.to_rfc3339(),
                 updated_at: now.to_rfc3339(),
                 mark: MarkSummary {
-                    earned: 0,
+                    earned: 0.0,
                     total: allocator.total_value,
                 },
                 is_practice: updated.is_practice,
@@ -720,7 +720,7 @@ async fn grade_submission(
 
     // Apply late cap if applicable
     let is_late_now = submission.created_at > assignment.due_date;
-    let mut _late_capped_to: Option<i64> = None;
+    let mut _late_capped_to: Option<f32> = None;
     if is_late_now && config.marking.late.allow_late_submissions {
         if within_late_window(
             submission.created_at,
@@ -1226,7 +1226,7 @@ pub async fn submit_assignment(
                 SubmissionStatus::FailedUpload,
                 Some("Disallowed code patterns detected".into()),
                 Some(WsMark {
-                    earned: 0,
+                    earned: 0.0,
                     total: response.mark.total,
                 }),
             )
@@ -1265,8 +1265,8 @@ pub async fn submit_assignment(
         assignment_id,
         claims.sub,
         attempt,
-        0,
-        0,
+        0.0,
+        0.0,
         is_practice,
         &file_name,
         &file_hash,
